added makefile

+++ MLearning/Clojure/earth-mover-distance/c_code/emd.h
@@ -24,7 +24,7 @@
 
 /*****************************************************************************/
 /* feature_t SHOULD BE MODIFIED BY THE USER TO REFLECT THE FEATURE TYPE      */
-typedef int feature_t;
+typedef double feature_t;
 /*****************************************************************************/
